+++ src/scirex/torch/sciml/data/__init__.py
@@ -5,6 +5,6 @@ This package provides tools for generating synthetic datasets for training
 Fourier Neural Operator models on various PDE problems.
 """
 
-from .generate import DataGenerator
+from .generate_sr import DataGenerator
 
 __all__ = ['DataGenerator']
