+++ src/scirex/torch/sciml/data/generate_sr.py
